Add support for extensions to configure lease coordination table.
