+++ data-prepper-plugins/kinesis-source/build.gradle
@@ -29,6 +29,10 @@ dependencies {
     testImplementation platform('org.junit:junit-bom:5.9.1')
     testImplementation 'org.junit.jupiter:junit-jupiter'
     testImplementation project(':data-prepper-test-event')
+    testImplementation project(':data-prepper-core')
+    testImplementation project(':data-prepper-plugin-framework')
+    testImplementation project(':data-prepper-pipeline-parser')
+    testImplementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml'
 }
 
 jacocoTestCoverageVerification {
