+++ data-prepper-plugins/kinesis-source/src/main/java/org/opensearch/dataprepper/plugins/kinesis/extension/KinesisLeaseConfig.java
@@ -1,0 +1,10 @@
+package org.opensearch.dataprepper.plugins.kinesis.extension;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+
+@Getter
+public class KinesisLeaseConfig {
+    @JsonProperty("lease_coordination_table")
+    private String leaseCoordinationTable;
+}
