+++ data-prepper-plugins/kinesis-source/src/main/java/org/opensearch/dataprepper/plugins/kinesis/extension/KinesisLeaseConfigExtension.java
@@ -1,0 +1,21 @@
+package org.opensearch.dataprepper.plugins.kinesis.extension;
+
+import org.opensearch.dataprepper.model.annotations.DataPrepperExtensionPlugin;
+import org.opensearch.dataprepper.model.annotations.DataPrepperPluginConstructor;
+import org.opensearch.dataprepper.model.plugin.ExtensionPlugin;
+import org.opensearch.dataprepper.model.plugin.ExtensionPoints;
+
+@DataPrepperExtensionPlugin(modelType = KinesisLeaseConfig.class, rootKeyJsonPath = "/kinesis_lease_config", allowInPipelineConfigurations = true)
+public class KinesisLeaseConfigExtension implements ExtensionPlugin {
+
+    private KinesisLeaseConfigSupplier kinesisLeaseConfigSupplier;
+    @DataPrepperPluginConstructor
+    public KinesisLeaseConfigExtension(final KinesisLeaseConfig kinesisLeaseConfig) {
+        this.kinesisLeaseConfigSupplier = new DefaultKinesisLeaseConfigSupplier(kinesisLeaseConfig);
+    }
+
+    @Override
+    public void apply(final ExtensionPoints extensionPoints) {
+        extensionPoints.addExtensionProvider(new KinesisLeaseConfigProvider(this.kinesisLeaseConfigSupplier));
+    }
+}
