+++ data-prepper-plugins/kinesis-source/src/main/java/org/opensearch/dataprepper/plugins/kinesis/extension/KinesisLeaseConfigSupplier.java
@@ -1,0 +1,10 @@
+package org.opensearch.dataprepper.plugins.kinesis.extension;
+
+import java.util.Optional;
+
+public interface KinesisLeaseConfigSupplier {
+
+    default Optional<KinesisLeaseConfig> getKinesisExtensionLeaseConfig() {
+        return Optional.empty();
+    }
+}
