--- data-prepper-plugins/kinesis-source/src/main/java/org/opensearch/dataprepper/plugins/source/kinesis/configuration/AwsAuthenticationConfig.java
@@ -1,87 +1,0 @@
-package org.opensearch.dataprepper.plugins.source.kinesis.configuration;
-
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.validation.constraints.Size;
-import software.amazon.awssdk.arns.Arn;
-import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
-import software.amazon.awssdk.auth.credentials.DefaultCredentialsProvider;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.sts.StsClient;
-import software.amazon.awssdk.services.sts.auth.StsAssumeRoleCredentialsProvider;
-import software.amazon.awssdk.services.sts.model.AssumeRoleRequest;
-
-import java.util.Map;
-import java.util.UUID;
-
-public class AwsAuthenticationConfig {
-    private static final String AWS_IAM_ROLE = "role";
-    private static final String AWS_IAM = "iam";
-
-    @JsonProperty("region")
-    @Size(min = 1, message = "Region cannot be empty string")
-    private String awsRegion;
-
-    @JsonProperty("sts_role_arn")
-    @Size(min = 20, max = 2048, message = "awsStsRoleArn length should be between 1 and 2048 characters")
-    private String awsStsRoleArn;
-
-    @JsonProperty("sts_external_id")
-    @Size(min = 2, max = 1224, message = "awsStsExternalId length should be between 2 and 1224 characters")
-    private String awsStsExternalId;
-
-    @JsonProperty("sts_header_overrides")
-    @Size(max = 5, message = "sts_header_overrides supports a maximum of 5 headers to override")
-    private Map<String, String> awsStsHeaderOverrides;
-
-    public Region getAwsRegion() {
-        return awsRegion != null ? Region.of(awsRegion) : null;
-    }
-
-    public String getAwsStsRoleArn() {
-        return awsStsRoleArn;
-    }
-
-    public String getAwsStsExternalId() {
-        return awsStsExternalId;
-    }
-
-    public Map<String, String> getAwsStsHeaderOverrides() {
-        return awsStsHeaderOverrides;
-    }
-
-    public AwsCredentialsProvider authenticateAwsConfiguration() {
-
-        final AwsCredentialsProvider awsCredentialsProvider;
-        if (awsStsRoleArn != null && !awsStsRoleArn.isEmpty()) {
-            try {
-                Arn.fromString(awsStsRoleArn);
-            } catch (final Exception e) {
-                throw new IllegalArgumentException("Invalid ARN format for awsStsRoleArn");
-            }
-
-            final StsClient stsClient = StsClient.builder().region(getAwsRegion()).build();
-
-            AssumeRoleRequest.Builder assumeRoleRequestBuilder = AssumeRoleRequest.builder()
-                    .roleSessionName("GeoIP-Processor-" + UUID.randomUUID()).roleArn(awsStsRoleArn);
-
-            if (awsStsHeaderOverrides != null && !awsStsHeaderOverrides.isEmpty()) {
-                assumeRoleRequestBuilder = assumeRoleRequestBuilder.overrideConfiguration(
-                        configuration -> awsStsHeaderOverrides.forEach(configuration::putHeader));
-            }
-
-            awsCredentialsProvider = StsAssumeRoleCredentialsProvider.builder()
-                    .stsClient(stsClient)
-                    .refreshRequest(assumeRoleRequestBuilder.build())
-                    .build();
-
-        } else {
-            awsCredentialsProvider = DefaultCredentialsProvider.create();
-        }
-        return awsCredentialsProvider;
-    }
-}
